[Pytorch Delegated Backend] Save function name in debug info
Summary:
This diff introduces function name to InlinedCallStack.
Since we are using InlinedCallStack for debug information in lite
interpreter as well as delegate backends, where InlinedCallStack cannot
be constructed from model source code, we need to save function name.
In the absence of function name Function* is used to get name of the
function. This is when JIT compiles code at runtime.
When that is not possible, this diff introduces a way to obtain function
name.

Test Plan:
test_backend
test_cs_debug_info_serialization

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 4faa2c287ae07b9cb33898ff150b4da257a855e0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57481